Create dev branch to build+push dev containers
This commit creates a dev branch with its own publish-container workflow. The new workflow will build
and push a "latest.dev"-tagged container on pushes or merges to the dev branch. This should allow
me to keep the main branch a little cleaner, and do the meat of the development in the dev branch.