Fix a few floating bugs
This commit fixes two bugs -- First, instantiation of the scheduler was moved into main to prevent multiple
app instances that caused duplication of scheduled tasks. Second, instantiation of the app itself was moved
into main to prevent the "blueprint already registered" bug that had been plaguing me.

In addition to these, I snuck in a few more comments, turned off an extraneous print statement, and added a
default "base path" to the X-OSDF-Authorization header for cases when the namespace has no base path in the
namespaces JSON. Non-related updates in the same commit, naughty naughty.