Switch collector to default
Switching the collector to the default init so that I can patch the various deployments
with the env var "_condor_COLLECTOR_HOST" that point to dev/production instances of the
collector. HTCondor will set the default collector to whatever the value of COLLECTOR_HOST
is, and _condor_COLLECTOR_HOST will override that value.