Fix header dict capitalization and add headers
Several small updates in this commit:
First, in following RFC 8941, which pertains to structured HTTP headers,
dict keys should be entirely lowercase.

Second, the OSDF Client relied on a few more pieces of information from
the namespaces.json that the Director didn't provide. Whether the
OSDF Client should be reworked to not need the extra info (eg ReadHTTPS
and UseTokenOnRead), or should the Director be configured to provide
that information remains to be seen. For now, the Director will supply
it.

Finally, the app was configured to get the namespaces.json and advertise
every minute, a vestige from debugging. The interval was changed to a
more reasonable 10 minutes.