Move base-path to x-osdf-token-generation
Prior to fully understanding why a base path was needed in some cases, I had put the base-path parameter
in the x-osdf-authorization header, and if no base path was provided by the namespace add, I made it the
same as the namespace path. Both of these were wrong. Now the base path is in the token generation header,
and if no base path is provided for the namespace, none is set.